Доработки по типу данных BigDecimal

+++ src/main/java/org/example/ref_calc/dto/CalculationRequestDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -12,7 +13,7 @@ import java.time.LocalDate;
 public class CalculationRequestDto {
     private LocalDate beginDate;
     private LocalDate endDate;
-    private double interestRate;
+    private BigDecimal interestRate;
     private double amount;
     private LocalDate paymentDate;
 }
