+++ src/main/java/org/example/ref_calc/dto/CalculationResponseDto.java
@@ -3,12 +3,13 @@ package org.example.ref_calc.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Setter
 public class CalculationResponseDto {
-    private double liability;
+    private BigDecimal liability;
     private List<PaymentDto> payments;
     private List<InterestExpenseDto> interestExpenses;
 //    private List<LiabilityDto> liabilities;
